+++ tests/contrib/print_proxy/test_mapfish_print.py
@@ -2,7 +2,7 @@
 import os
 import json
 import codecs
-from pyramid_oereb.contrib.print_proxy.mapfish_print import Renderer
+from pyramid_oereb.contrib.print_proxy.mapfish_print.mapfish_print import Renderer
 from tests.renderer import DummyRenderInfo
 from pyramid_oereb.contrib.print_proxy.mapfish_print.toc_pages import TocPages
