+++ tests/core/renderer/xml/test_legend_entry.py
@@ -1,10 +1,10 @@
-from tests.renderer import DummyRenderInfo
-from tests.renderer.xml import xml_templates
+from tests.core.renderer import DummyRenderInfo
+from tests.core.renderer.xml import xml_templates
 from pyramid_oereb.core.records.image import ImageRecord
 from pyramid_oereb.core.records.theme import ThemeRecord
 from pyramid_oereb.core.records.view_service import LegendEntryRecord
 from pyramid_oereb.core.renderer.extract.xml_ import Renderer
-from pyramid_oereb.core.views import Parameter
+from pyramid_oereb.core.views.webservice import Parameter
 from tests.mockrequest import MockRequest
 
 template = xml_templates().get_template('legend_entry.xml')
