+++ tests/core/webservice/test_getegrid.py
@@ -13,7 +13,7 @@ from pyramid_oereb.core.records.real_estate import RealEstateRecord
 from pyramid_oereb.core.records.view_service import ViewServiceRecord
 from tests import schema_json_extract, pyramid_oereb_test_config
 from tests.mockrequest import MockRequest
-from pyramid_oereb.core.views import PlrWebservice, Parameter
+from pyramid_oereb.core.views.webservice import PlrWebservice, Parameter
 
 
 def test_getegrid_coord_missing_parameter():
