+++ tests/core/webservice/test_getversions.py
@@ -4,7 +4,7 @@ import json
 from jsonschema import Draft4Validator
 from tests import schema_json_versions, pyramid_oereb_test_config
 from tests.mockrequest import MockRequest
-from pyramid_oereb.core.views import PlrWebservice
+from pyramid_oereb.core.views.webservice import PlrWebservice
 
 
 def test_getversions_json():
