+++ tests/mockrequest.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from pyramid.testing import DummyRequest
 
-from pyramid_oereb.core.views import Parameter
+from pyramid_oereb.core.views.webservice import Parameter
 
 
 class MockParameter(Parameter):
